requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.34.0
python-docx==1.1.0
lxml==5.2.1
gunicorn==21.2.0
